Bump version and change created field type to float
Update __version__ to 0.1.11 and modify the ChatCompletionModel created field from int to float to allow greater precision. This ensures compatibility with timestamps requiring fractional seconds.

+++ magic_llm/model/ModelChatStream.py
@@ -41,7 +41,7 @@ class UsageModel(BaseModel):
 class ChatCompletionModel(BaseModel):
     id: str
     object: Optional['str'] = 'chat.completion.chunk'
-    created: Optional[int] = None
+    created: Optional[float] = None
     model: str
     system_fingerprint: Optional[Any] = None
     choices: List[ChoiceModel]
